python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.76
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
gunicorn==21.2.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
huggingface-hub==0.17.3 